+++ Gemfile
@@ -1,35 +1,5 @@
 source 'https://rubygems.org'
 
-gem 'sinatra'
-gem 'tilt', '~> 1.4.1'
-gem 'tilt-jbuilder', require: 'sinatra/jbuilder'
+gemspec
 
-group :production do
-  gem 'foreman'
-  gem 'unicorn'
-end
-
-gem 'httparty'
-gem 'honeybadger'
-gem 'peach'
-
-gem 'endpoint_base', github: 'spree/endpoint_base'
-
-group :development, :test do
-  gem 'shotgun'
-  gem 'pry'
-  gem 'pry-byebug'
-  gem 'simplecov', require: false
-  gem 'dotenv'
-end
-
-group :test do
-  gem 'vcr'
-  gem 'rspec'
-  gem 'webmock'
-  gem 'guard-rspec'
-  gem 'terminal-notifier-guard'
-  gem 'rb-fsevent', '~> 0.9.1'
-  gem 'rack-test'
-  gem 'timecop'
-end
+gem 'endpoint_base', git: 'https://github.com/Follain/endpoint_base'
