+++ config.ru
@@ -2,5 +2,5 @@ require 'rubygems'
 require 'bundler'
 
 Bundler.require(:default)
-require './vend_endpoint.rb'
+require './lib/vend_endpoint.rb'
 run VendEndpoint
